[ refactor ] : core-model conflict resolve
