+++ core-model/src/main/java/com/ddd/ansayo/core_model/common/Response.kt
@@ -1,4 +1,4 @@
-package com.ddd.ansayo.domain.model.common
+package com.ddd.ansayo.core_model.common
 
 sealed class Response<out T> {
     data class Success<T>(val data: T) : Response<T>()
