+++ core-model/src/main/java/com/ddd/ansayo/core_model/course/UploadImageUrlEntity.kt
@@ -1,0 +1,5 @@
+package com.ddd.ansayo.core_model.course
+
+data class UploadImageUrlEntity(
+    val uploadUrl: String
+)
