+++ data/src/main/java/com/ddd/ansayo/data/datasource/course/CourseRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.ddd.ansayo.data.datasource.course
 
-import com.ddd.ansayo.data.model.course.UploadImageUrlData
+import com.ddd.ansayo.core_model.course.UploadImageUrlEntity
 
 interface CourseRemoteDataSource {
 
-    suspend fun getUploadImageUrl(code: String, fileName: String): UploadImageUrlData
+    suspend fun getUploadImageUrl(code: String, fileName: String): UploadImageUrlEntity
 }
