--- data/src/main/java/com/ddd/ansayo/data/model/common/DataModel.kt
@@ -1,3 +1,0 @@
-package com.ddd.ansayo.data.model.common
-
-interface DataModel
