--- data/src/main/java/com/ddd/ansayo/data/model/course/UploadImageUrlData.kt
@@ -1,7 +1,0 @@
-package com.ddd.ansayo.data.model.course
-
-import com.ddd.ansayo.data.model.common.DataModel
-
-data class UploadImageUrlData(
-    val uploadUrl: String
-) : DataModel
