+++ data/src/main/java/com/ddd/ansayo/data/repository/course/CourseRepositoryImpl.kt
@@ -1,9 +1,7 @@
 package com.ddd.ansayo.data.repository.course
 
+import com.ddd.ansayo.core_model.course.UploadImageUrlEntity
 import com.ddd.ansayo.data.datasource.course.CourseRemoteDataSource
-import com.ddd.ansayo.data.util.toResponse
-import com.ddd.ansayo.domain.entity.course.UploadImageUrlEntity
-import com.ddd.ansayo.domain.model.common.Response
 import com.ddd.ansayo.domain.repository.CourseRepository
 import javax.inject.Inject
 
@@ -14,7 +12,7 @@ class CourseRepositoryImpl @Inject constructor(
     override suspend fun getUploadImageUrl(
         code: String,
         fileName: String
-    ): Response<UploadImageUrlEntity> {
-        return courseRemoteDataSource.getUploadImageUrl(code, fileName).toResponse()
+    ): UploadImageUrlEntity {
+        return courseRemoteDataSource.getUploadImageUrl(code, fileName)
     }
 }
