--- domain/src/main/java/com/ddd/ansayo/domain/entity/common/EntityModel.kt
@@ -1,3 +1,0 @@
-package com.ddd.ansayo.domain.entity.common
-
-interface EntityModel
