--- domain/src/main/java/com/ddd/ansayo/domain/entity/course/UploadImageUrlEntity.kt
@@ -1,7 +1,0 @@
-package com.ddd.ansayo.domain.entity.course
-
-import com.ddd.ansayo.domain.entity.common.EntityModel
-
-data class UploadImageUrlEntity(
-    val uploadUrl: String
-) : EntityModel
