+++ domain/src/main/java/com/ddd/ansayo/domain/usecase/course/GetImageUploadUrlUseCase.kt
@@ -1,8 +1,9 @@
 package com.ddd.ansayo.domain.usecase.course
 
-import com.ddd.ansayo.domain.entity.course.UploadImageUrlEntity
-import com.ddd.ansayo.domain.model.common.Response
+import com.ddd.ansayo.core_model.course.UploadImageUrlEntity
+import com.ddd.ansayo.core_model.common.Response
 import com.ddd.ansayo.domain.repository.CourseRepository
+import com.ddd.ansayo.domain.util.toResponse
 import javax.inject.Inject
 
 class GetImageUploadUrlUseCase @Inject constructor(
@@ -10,6 +11,6 @@ class GetImageUploadUrlUseCase @Inject constructor(
 ) {
 
     suspend operator fun invoke(code: String, fileName: String): Response<UploadImageUrlEntity> {
-        return courseRepository.getUploadImageUrl(code, fileName)
+        return courseRepository.getUploadImageUrl(code, fileName).toResponse()
     }
 }
