+++ remote/src/main/java/com/ddd/ansayo/remote/datasource/CourseRemoteDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.ddd.ansayo.remote.datasource
 
+import com.ddd.ansayo.core_model.course.UploadImageUrlEntity
 import com.ddd.ansayo.data.datasource.course.CourseRemoteDataSource
-import com.ddd.ansayo.data.model.course.UploadImageUrlData
 import com.ddd.ansayo.remote.service.CourseService
 import com.skydoves.sandwich.getOrThrow
 import javax.inject.Inject
@@ -13,7 +13,7 @@ class CourseRemoteDataSourceImpl @Inject constructor(
     override suspend fun getUploadImageUrl(
         code: String,
         fileName: String
-    ): UploadImageUrlData {
+    ): UploadImageUrlEntity {
         return courseService.getImageUploadUrl(code, fileName).getOrThrow()
     }
 }
