+++ remote/src/main/java/com/ddd/ansayo/remote/service/CourseService.kt
@@ -1,6 +1,6 @@
 package com.ddd.ansayo.remote.service
 
-import com.ddd.ansayo.data.model.course.UploadImageUrlData
+import com.ddd.ansayo.core_model.course.UploadImageUrlEntity
 import com.skydoves.sandwich.ApiResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -11,5 +11,5 @@ interface CourseService {
     suspend fun getImageUploadUrl(
         @Query("code") code: String,
         @Query("name") name: String
-    ): ApiResponse<UploadImageUrlData>
+    ): ApiResponse<UploadImageUrlEntity>
 }
